init
init

+++ resources/views/products/products.blade.php
@@ -2,6 +2,7 @@
     $p = new App\Helpers\Picker;
     $info = new App\Helpers\Info;
     $r = new App\Helpers\Role;
+    $f = new App\Helpers\Filter;
 ?>
 @extends('../nav')
 
@@ -35,33 +36,38 @@
             </div>
 
                 @foreach($record->products as $product)
-                <div class="col-4 col-sm-2 text-center">
-                    <a href="/products/show/{{ $product->id }}">
-                      <img src="{{ $product->img }}" class="rounded product-icon"></a>
-                    @if($product->fs)
-                        <span class="badge text-white water product-badge"><i class="fa fa-shower" aria-hidden="true"></i></span>
-                    @else
-                        @if($p->fresh($product->id))
-                            <span class="badge text-white leaf product-badge"><i class="fa fa-leaf" aria-hidden="true"></i></span>
-                        @endif
-                    @endif
-   
-                    <p class="text-center">
-                        <strong>¥{{ $product->quota }}</strong><br>
-                        <strong>{{ $product->name }}</strong>
-                        @if($r->issuer())
-                            <br>
+                    @if($f->onLine($product))
 
-                            <a href="/products/slide/{{ $product->id }}" class="badge badge-warning"><i class="fa fa-star" aria-hidden="true"></i> 明星</a>
-                            @if($product->show)
-                                <a href="/products/off/{{ $product->id }}" class="badge badge-warning"><i class="fa fa-fire" aria-hidden="true"></i> 下架</a>
+                        <div class="col-4 col-sm-2 text-center">
+                            <a href="/products/show/{{ $product->id }}">
+                              <img src="{{ $product->img }}" class="rounded product-icon"></a>
+                            @if($product->fs)
+                                <span class="badge text-white water product-badge"><i class="fa fa-shower" aria-hidden="true"></i></span>
                             @else
-                                <a href="/products/on/{{ $product->id }}" class="badge badge-success"><i class="fa fa-leaf" aria-hidden="true"></i> 上架</a>
+                                @if($p->fresh($product->id))
+                                    <span class="badge text-white leaf product-badge"><i class="fa fa-leaf" aria-hidden="true"></i></span>
+                                @endif
                             @endif
-                            <a href="/products/delete/{{ $product->id }}" class="badge badge-danger"><i class="fa fa-flash" aria-hidden="true"></i> 删除!</a>
-                        @endif
-                    </p>
-                </div>
+           
+                            <p class="text-center">
+                                <strong>¥{{ $product->quota }}</strong><br>
+                                <strong>{{ $product->name }}</strong>
+
+                                @if($r->issuer())
+                                    <br>({{ $product->org->name }}-{{ $product->type->val }})
+                                    <br>
+                                    <a href="/products/slide/{{ $product->id }}" class="badge badge-warning"><i class="fa fa-star" aria-hidden="true"></i> 明星</a>
+                                    @if($product->show)
+                                        <a href="/products/off/{{ $product->id }}" class="badge badge-warning"><i class="fa fa-fire" aria-hidden="true"></i> 下架</a>
+                                    @else
+                                        <a href="/products/on/{{ $product->id }}" class="badge badge-success"><i class="fa fa-leaf" aria-hidden="true"></i> 上架</a>
+                                    @endif
+                                    <a href="/products/delete/{{ $product->id }}" class="badge badge-danger"><i class="fa fa-flash" aria-hidden="true"></i> 删除!</a>
+                                @endif
+                            </p>
+                        </div>
+                        
+                    @endif
                 @endforeach
 
             @endforeach
